Refactored unit tests
The current structure of unit tests became too chaotic as more unit
tests need to be added in the future. This patch groups some of the
tests by the corresponding decorator to make categorization of new tests
a bit easier.
